fix(revenue-ui): change RevenueGrowthPercentage type from double to decimal for consistency

+++ server/Application/Common/DTOs/Department/ResponseDeparmentRevenueStatisticsDTO.cs
@@ -4,5 +4,5 @@ public class ResponseDeparmentRevenueStatisticsDTO
     public string Name { get; set; } = string.Empty;
     public int TotalAppointments { get; set; }
     public decimal Revenue { get; set; }
-    public double? RevenueGrowthPercentage { get; set; }
+    public decimal? RevenueGrowthPercentage { get; set; }
 }
